RD-4423: Refactored components
