+++ widgets/tokens/src/TextEllipsis.styles.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const EllipsisWrapper = styled.div`
+    max-width: 300px;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+`;
