+++ widgets/tokens/src/TextEllipsis.tsx
@@ -1,11 +1,4 @@
-import styled from 'styled-components';
-
-const EllipsisWrapper = styled.div`
-    max-width: 300px;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-`;
+import { EllipsisWrapper } from './TextEllipsis.styles';
 
 interface TextEllipsisProps {
     content: string | null;
