+++ widgets/tokens/src/TokensTableHeader/CreateTokenModal.types.ts
@@ -1,0 +1,9 @@
+export interface ReceivedToken {
+    id: string;
+    description: string;
+    expiration_date: null;
+    last_used: null;
+    username: string;
+    value: string;
+    role: string;
+}
